turbo-tasks: split scope.rs into scope_bounded and scope_unbounded
Review feedback: the two scope APIs were interleaved in one file and the
unbounded one named its handle `Spawner` to avoid colliding with the
bounded `Scope`. Splitting by module lets both handles be called `Scope`
and disambiguates by path instead of by suffix.

- `scope_bounded`: `Scope`, `scope_and_block`. Task count known up front;
  the closure runs once and spawns closures onto the scope.
- `scope_unbounded`: `Scope` (was `Spawner`), `scope_unbounded`,
  `scope_unbounded_with`. Work is discovered as it runs; the closure *is*
  the per-item body.

Module docs now explain why the unbounded API isn't shaped like
`std::thread::scope`: there is no single call to hand a scope object to
because `run` is invoked per item, so the seed set arrives as `initial`
and early termination goes through `ControlFlow::Break` — aborting has to
reach the shared queue, not just return from one invocation.

Pure code motion plus that rename: no logic changed, all 30 tests move
with their modules (10 bounded, 20 unbounded) and still pass.
