+++ turbopack/crates/turbo-tasks/src/lib.rs
@@ -47,7 +47,8 @@ mod priority_runner;
 mod read_options;
 mod read_ref;
 pub mod registry;
-pub mod scope;
+pub mod scope_bounded;
+pub mod scope_unbounded;
 mod serialization_invalidation;
 pub mod small_duration;
 mod spawn;
